Clean test runner up (#8485)
* bump scale-codec in test runner

* refactor config

* Update test-utils/test-runner/Cargo.toml


* bump cargo.lock

* add reasonable defaults
