+++ test-utils/test-runner/src/lib.rs
@@ -228,7 +228,7 @@
 
 use manual_seal::consensus::ConsensusDataProvider;
 use sc_executor::NativeExecutionDispatch;
-use sc_service::{Configuration, TFullBackend, TFullClient, TaskManager};
+use sc_service::{Configuration, TFullBackend, TFullClient, TaskManager, TaskExecutor};
 use sp_api::{ConstructRuntimeApi, TransactionFor};
 use sp_consensus::{BlockImport, SelectChain};
 use sp_inherents::InherentDataProviders;
@@ -242,6 +242,7 @@ mod host_functions;
 
 pub use host_functions::*;
 pub use node::*;
+pub use utils::*;
 
 /// Wrapper trait for concrete type required by this testing framework.
 pub trait ChainInfo: Sized {
@@ -278,6 +279,9 @@ pub trait ChainInfo: Sized {
 
 	/// Signed extras, this function is caled in an externalities provided environment.
 	fn signed_extras(from: <Self::Runtime as frame_system::Config>::AccountId) -> Self::SignedExtras;
+
+	/// config factory
+	fn config(task_executor: TaskExecutor) -> Configuration;
 
 	/// Attempt to create client parts, including block import,
 	/// select chain strategy and consensus data provider.
